+++ changelog.md
@@ -5,7 +5,7 @@
 > \[!IMPORTANT\]
 > This version is not released yet and is under active development.
 
-- \[mpm\] Add new command to export installed packages to a SPDX file. Refs {issue}`936`.
+- \[mpm\] Add new command to export installed packages to a SBOM file in SPDX or CycloneDX standard. Closes {issue}`936`.
 
 ## [5.17.0 (2024-07-07)](https://github.com/kdeldycke/meta-package-manager/compare/v5.16.0...v5.17.0)